+++ detekt-test-utils/api/detekt-test-utils.api
@@ -27,11 +27,6 @@ public final class dev/detekt/test/utils/KotlinCoreEnvironmentWrapper : java/lan
 	public final fun getEnv ()Ldev/detekt/test/utils/KotlinEnvironmentContainer;
 }
 
-public final class dev/detekt/test/utils/KotlinCoreEnvironmentWrapperKt {
-	public static final fun createEnvironment (Ljava/util/List;Ljava/util/List;)Ldev/detekt/test/utils/KotlinCoreEnvironmentWrapper;
-	public static synthetic fun createEnvironment$default (Ljava/util/List;Ljava/util/List;ILjava/lang/Object;)Ldev/detekt/test/utils/KotlinCoreEnvironmentWrapper;
-}
-
 public final class dev/detekt/test/utils/KotlinEnvironmentContainer {
 	public fun <init> (Lcom/intellij/openapi/project/Project;Lorg/jetbrains/kotlin/config/CompilerConfiguration;)V
 	public final fun getConfiguration ()Lorg/jetbrains/kotlin/config/CompilerConfiguration;
